+++ HISTORY.rst
@@ -160,3 +160,4 @@ Changes
 * Vec3 and Group upper and flip properties implemented
 * DefGrad, VelGrad and Stress rotate method accepts also rotation matrix
 * velgrad method added to DefGrad to calculate matrix logarithm
+* StereoGrid has new methods max, min, max_at, min_at